Remove CTEUtils
